Running pyupgrade+black+isort on existing sources
We might have an issue with pyupgrade and python2. It replaces %s by .format (https://github.com/asottile/pyupgrade/pull/41/files)

```
-        method_name = '_send_invoice_%s' % (sending_method,)
+        method_name = '_send_invoice_{}'.format(sending_method)
```

Which may be a problem if the interpolation was mixing bytes and unicode (which
works by luck with %s but not with .format()). Demonstration:

```
>>> '%s' % (u'é',)
u'\xe9'
>>> '{}'.format(u'é')
Traceback (most recent call last):
  File "<console>", line 1, in <module>
  UnicodeEncodeError: 'acan't encode character u'\xe9' in position 0: ordinal not in range(128)
```