DRAFT base reserve rule on package
Draft of a rule based on packages. We maybe want to keep the
original packaging qty rule, not replace it, but the new test pass
(others would need to be adapted to use stock.quant.package records as
the new rule use them).

Not sure about the interaction with package/lot/owner if a rule set a
QuantProperties asking one but the _update_reserved_quantity method was
called with one already.